Works/FreakinHotelBot/requirements.txt
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
peewee==3.16.3
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.6
